pyproject.toml
[project]
name = "flowgenius"
version = "0.5.15"
description = "AI-assisted learning assistant that eliminates research paralysis"
readme = "README.md"
authors = [
    { name = "bbaserdem", email = "baserdemb@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "langchain-core>=0.3.0",
    "click>=8.0.0",
    "platformdirs>=4.0.0",
    "pydantic-settings>=2.0.0",
    "ruamel.yaml>=0.18.0",
    "textual>=0.80.0",
    "questionary>=2.1.0",
]

[project.scripts]
flowgenius = "flowgenius:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.1",
    {include-group = "test"}
]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
